// src/lib/client/models/Religion.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */

export type Religion = {
  readonly uuid: string;
  /**
   * The name of the religion or belief system.
   */
  name: string;
  /**
   * Brief description of the religion's core tenets.
   */
  description?: string | null;
};